+++ src/pages/article-copy-tool/components/HeaderArea/components/AdditionalConfig/components/ConfigSwitchButton.vue
