+++ src/pages/article-copy-tool/components/HeaderArea/components/HelpButton.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import {HelpCircle} from "@vicons/ionicons5";
-import useIntroducer from "../../../assets/ts/article-copy-tool/useIntroducer";
+import useIntroducer from "../../../../../assets/ts/article-copy-tool/useIntroducer";
 
 const {introduce} = useIntroducer();
 </script>
