+++ src/pages/article-copy-tool/components/HeaderArea/components/ThemeConfig.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import {useStore} from "../../../store/useStore";
+import {useStore} from "../../../../../store/useStore";
 import {Moon, Sunny} from "@vicons/ionicons5";
 
 const store = useStore();
