+++ src/pages/article-copy-tool/components/InputAndOutputArea/components/OutputFuncButtons.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import {useStore} from "../../../store/useStore";
+import {useStore} from "../../../../../store/useStore";
 import {useMessage} from "naive-ui";
 import useClipboard from "vue-clipboard3";
